remove DecodeV1 and EncodeV1 in the compiler
