+++ compiler/daml-lf-proto-decode/BUILD.bazel
@@ -8,7 +8,6 @@ da_haskell_library(
     srcs = [
         "src/DA/Daml/LF/Proto3/Archive/Decode.hs",
         "src/DA/Daml/LF/Proto3/Decode.hs",
-        "src/DA/Daml/LF/Proto3/DecodeV1.hs",
         "src/DA/Daml/LF/Proto3/DecodeV2.hs",
     ],
     hackage_deps = [
