+++ compiler/daml-lf-proto-encode/BUILD.bazel
@@ -8,7 +8,6 @@ da_haskell_library(
     srcs = [
         "src/DA/Daml/LF/Proto3/Archive/Encode.hs",
         "src/DA/Daml/LF/Proto3/Encode.hs",
-        "src/DA/Daml/LF/Proto3/EncodeV1.hs",
         "src/DA/Daml/LF/Proto3/EncodeV2.hs",
     ],
     hackage_deps = [
