+++ compiler/daml-lf-proto-util/src/DA/Daml/LF/Proto3/Util.hs
@@ -21,7 +21,7 @@ import qualified Data.Text.Lazy as TL
 import           GHC.Generics
 import qualified Numeric
 
-import qualified Com.Daml.DamlLfDev.DamlLf1 as P
+import qualified Com.Daml.DamlLfDev.DamlLf2 as P
 
 class EitherLike a b e where
     toEither :: e -> Either a b
