+++ compiler/damlc/tests/src/DA/Test/DamlcIntegration.hs
@@ -31,7 +31,7 @@ import           Control.Exception.Extra
 import           Control.Monad
 import           Control.Monad.IO.Class
 import           DA.Daml.LF.PrettyScenario (prettyScenarioError, prettyScenarioResult)
-import           DA.Daml.LF.Proto3.EncodeV1
+import           DA.Daml.LF.Proto3.EncodeV2
 import qualified DA.Daml.LF.Proto3.Archive.Encode as Archive
 import           DA.Pretty hiding (first)
 import qualified DA.Daml.LF.ScenarioServiceClient as SS
